chore(log): move log to shared libraries
